feat: format statuses and annoations data field as string in events (#179)
This updates the `data` field on `Status` and `Annotation` to be sent as
a string on change events. Currently it gets sent as a []byte which
makes it harder to work with for consumers.

Some other little cleanup things:
- on renovate generate action check if it's a renovate PR before
checking out the code since forks don't have access to secrets
- fix the linting errors stopping #171 from being merged

---------

Signed-off-by: Nicole Hubbard <nhubbard@groq.com>